Refactor util -> util, etc; Removed circular dependency
