+++ util/chplenv/chpl_3p_libunwind_configs.py
@@ -4,8 +4,8 @@ import sys
 chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
-import third_party_utils, utils
 import chpl_platform
+import third_party_utils
 from utils import memoize
