+++ util/chplenv/third_party_utils.py
@@ -5,7 +5,8 @@ import sys
 chplenv_dir = os.path.dirname(__file__)
 sys.path.insert(0, os.path.abspath(chplenv_dir))
 
-import chpl_arch, chpl_compiler, chpl_locale_model, chpl_platform, utils
+import chpl_arch, chpl_compiler, chpl_locale_model, chpl_platform
+from etc import get_chpl_home, using_chapel_module
 from utils import memoize
 
 
@@ -17,14 +18,14 @@ def default_uniq_cfg_path():
     return '{0}-{1}-{2}'.format(chpl_platform.get('target'),
                                 chpl_compiler.get('target'),
                                 chpl_arch.get('target', map_to_compiler=True,
-                                              get_lcd=utils.using_chapel_module()))
+                                              get_lcd=using_chapel_module()))
 
 #
 # Returns the path to the packages install directory
 #
 @memoize
 def get_cfg_install_path(pkg, ucp=default_uniq_cfg_path()):
-    return os.path.join(utils.get_chpl_home(), 'third-party', pkg, 'install', ucp)
+    return os.path.join(get_chpl_home(), 'third-party', pkg, 'install', ucp)
 
 #
 # Return libraries and other options mentioned in the old_library and
@@ -48,7 +49,7 @@ def handle_la(la_path):
                     for tok in line.split('\'')[1].split():
                         # paths reflect built env; replace with $CHPL_HOME
                         pat = re.compile(r'^((-L\s*)?).*(/third-party/)')
-                        repl = r'\1' + utils.get_chpl_home() + r'\3'
+                        repl = r'\1' + get_chpl_home() + r'\3'
                         tok = pat.sub(repl, tok)
                         if tok.endswith('.la'):
                             args.extend(handle_la(tok))
